+++ source/mango/image/block_fxt1.cpp
@@ -19,7 +19,7 @@ namespace
         u32 getMode() const
         {
             // mode is stored in the 3 last bits of FXT block
-            return data[15] >> 5;
+            return u32(data[15] >> 5);
         }
     };
 
@@ -277,8 +277,8 @@ namespace
 
         if (!block.alpha)
         {
-            u32 bit01 = (block.indices[0] >> 1) & 1;
-            u32 bit33 = (block.indices[4] >> 1) & 1;
+            u32 bit01 = u32((block.indices[0] >> 1) & 1);
+            u32 bit33 = u32((block.indices[4] >> 1) & 1);
 
             u32 b0 = expand5to8(block.blue0);
             u32 g0 = expand5to8(u32(block.green0 << 1) | u32(block.lsb1 ^ bit01));
